Add thpool to oehost

+++ common/thpool.c
@@ -149,8 +149,8 @@ struct thpool_* thpool_init(int num_threads)
     }
 
     /* Make threads in pool */
-    thpool_p->threads =
-        (struct thread**)malloc(num_threads * sizeof(struct thread*));
+    thpool_p->threads = (struct thread**)malloc(
+        (unsigned long)num_threads * sizeof(struct thread*));
     if (thpool_p->threads == NULL)
     {
         err("thpool_init(): Could not allocate memory for threads\n");
